파일이동수정

+++ src/main/java/com/example/codebase/Service/FileService.java
@@ -266,7 +266,7 @@ public class FileService {
 
         File from = new File(rootPath + member_id + dir);
         File to = new File(rootPath+member_id+mv_dir+"/" +from.getName());
-        if(!to.isDirectory()&&from.exists()){
+        if(to.isDirectory()&&from.exists()&& !to.exists()){
             to.mkdir();
             FileUtils.copyDirectory(from,to);
             Files.walk(from.toPath())
@@ -275,7 +275,10 @@ public class FileService {
                     .forEach(File::delete);
                     }
         else if(to.isDirectory()) return new ResponseEntity<>("이미 폴더가 존재함",HttpStatus.OK);
-        else return new ResponseEntity<>("존재하지 않는 폴더",HttpStatus.OK);
+        else if(to.isFile()&&from.exists()&&!to.exists()){
+            FileUtils.moveFile(from,to);
+        }
+        else return new ResponseEntity<>("존재하지 않는 파일",HttpStatus.OK);
         return new ResponseEntity<>("파일 이동 완료",HttpStatus.OK);
     }
